+++ tests/testthat/test_Taglist.R
@@ -1,4 +1,3 @@
-context("test_Taglist")
 #* @skip