+++ tests/testthat/test_get_test_index.R
@@ -1,9 +1,4 @@
-context("get_test_index")
-
-
-
-
-setup({
+setup <- function(){
   tenv <- parent.env(environment())
   proj_old <- tryCatch(usethis::proj_get(), error = function(e) NULL)
   assign("proj_old", proj_old, tenv)
@@ -13,20 +8,22 @@ setup({
   dir <- find_testdata("test_pkg", must_exist = TRUE)
   fs::dir_copy(dir, proj_test)
   usethis::proj_set(proj_test)
-})
+}
 
 
-
-
-teardown({
+teardown <- function(){
   usethis::proj_set(proj_old)
   unlink(td, recursive = TRUE)
-})
+}
+
 
 
 
 
 test_that("extract_testthat_parse_data works as expected", {
+  setup()
+  on.exit(teardown())
+
   # works with text
   txt <- "test_that(\n\n'blubb')"
   x <- extract_testthat_parse_data(txt)
@@ -47,6 +44,9 @@ test_that("extract_testthat_parse_data works as expected", {
 
 
 test_that("extract_testthat_desc works as expected", {
+  setup()
+  on.exit(teardown())
+
   # works with text
   txt <- "test_that(\n\n'blubb')"
   x <- extract_testthat_desc(txt)
@@ -72,6 +72,9 @@ test_that("extract_testthat_desc works as expected", {
 
 
 test_that("collect_testthat_source_info works as expected", {
+  setup()
+  on.exit(teardown())
+
   src <-
     find_testdata("test_pkg", "tests", "testthat", "testthat_parse_cases.R")